v0.3.3: fix contract deploy metadata, premium refund, auto-register fallback
Bug fixes:
- Fix AccountInfo.molt u64 -> f64 deserialization (RPC returns float)
- Deploy premium (25 MOLT) refunded on b_put_transaction/commit_batch failures
- Deploy confirmation: 10 retries x 1s (was 5 x 500ms)
- Auto-register symbol fallback when deploy succeeds but registry entry missing
- Better from_init_data error logging for deploy debugging

Tests:
- test_contract_program_deploy_with_symbol_registry
- test_contract_program_deploy_failure_refunds_premium

Docs:
- SKILL.md: VPS state directory convention (/var/lib/moltchain/), deploy output docs
- README.md: update version to v0.3.3
- developers/contracts.html: expanded deploy section with all flags + info box

+++ cli/src/client.rs
@@ -172,7 +172,7 @@ pub struct StakingRewards {
 pub struct AccountInfo {
     pub pubkey: String,
     pub balance: u64,
-    pub molt: u64,
+    pub molt: f64,
     pub exists: bool,
     pub is_executable: bool,
     pub is_validator: bool,
